requirements.txt
requests
beautifulsoup4
tqdm
googlesearch-python
colorama
